requirements.txt
aiofiles==0.6.0
click==7.1.2
cmake==3.18.4.post1
cycler==0.10.0
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.63.0
h11==0.12.0
importlib-metadata==3.10.1
Jinja2==2.11.3
jjy==0.0.3.4
jsonpickle==2.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
Pillow==8.2.0
pydantic==1.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
starlette==0.13.6
tqdm==4.60.0
typing-extensions==3.7.4.3
uvicorn==0.13.4
zipp==3.4.1
